+++ examples/my-app/CHANGELOG.md
@@ -1,5 +1,16 @@
 # my-app
 
+## 1.40.0
+
+### Minor Changes
+
+- fix bugs
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/provider@1.41.0
+
 ## 1.39.0
 
 ### Minor Changes