+++ packages/multicall/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @0xsodium/multicall
 
+## 1.41.0
+
+### Minor Changes
+
+- fix bugs
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/abi@1.41.0
+  - @0xsodium/network@1.41.0
+  - @0xsodium/utils@1.41.0
+
 ## 1.40.0
 
 ### Minor Changes