+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @0xsodium/provider
 
+## 1.41.0
+
+### Minor Changes
+
+- fix bugs
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/transactions@1.41.0
+  - @0xsodium/wallet@1.41.0
+  - @0xsodium/abi@1.41.0
+  - @0xsodium/auth@1.41.0
+  - @0xsodium/config@1.41.0
+  - @0xsodium/graphquery@1.41.0
+  - @0xsodium/network@1.41.0
+  - @0xsodium/utils@1.41.0
+
 ## 1.40.0
 
 ### Minor Changes